# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Авито Питомец — API",
    "version": "1.0.0",
    "description": "Контракт фронт ↔ бэк.\n\nПРАВИЛА, ОБЯЗАТЕЛЬНЫЕ ДЛЯ ОБЕИХ СТОРОН:\n1. Клиент НЕ передаёт XP, уровень, прогресс задач, значения показателей. Только факт действия и его идентификатор.\n2. Все расчёты (decay, XP, уровень, настроение, стрик, право на награду) — на сервере.\n3. Порядок множителей: XP = базовый × множитель_настроения × множитель_стрика. Настроение берётся ПОСЛЕ изменения показателя. Округление вниз до целого.\n4. Показатели целые 0..100, потолок жёсткий: 82 + 30 = 100, остаток сгорает.\n5. Суточный потолок ухода обрезает последнее действие ЧАСТИЧНО (осталось 5 из 40 — начисляем 5, не 0).\n6. Сутки и стрик считаются по timezone аккаунта, а не по времени устройства.\n7. Все ответы в конверте { data, meta }.",
    "contact": { "name": "Фронтенд", "email": "frontend@avito-pet.local" }
  },
  "servers": [
    { "url": "https://api.avito-pet.local/api/v1", "description": "прод" },
    { "url": "http://localhost:8080/api/v1", "description": "локально через docker-compose" }
  ],
  "security": [ { "bearerAuth": [] } ],
...
  "paths": {
...
    "/pets": {
      "post": {
        "tags": ["pet"], "summary": "Создать питомца",
        "parameters": [ { "$ref": "#/components/parameters/IdempotencyKey" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["presetId"],
          "properties": {
            "presetId": { "$ref": "#/components/schemas/PresetId" },
            "name": { "type": "string", "maxLength": 16 }
          } } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/PetOk" },
          "409": { "description": "Питомец уже создан", "content": { "application/json": { "schema": {
            "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/pet": {
      "get": {
        "tags": ["pet"],
        "summary": "Состояние питомца",
        "description": "Сервер лениво досчитывает decay от pet.updatedAt до now и, если это первый вход за сутки, начисляет XP за вход. Крон для показателей не нужен.",
        "responses": {
          "200": { "$ref": "#/components/responses/PetOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Питомца ещё нет — вести на онбординг", "content": { "application/json": { "schema": {
            "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/pet/actions": {
      "post": {
        "tags": ["pet"],
        "summary": "Действие ухода",
        "description": "Единственный способ поднять показатель. Возвращает ПОЛНОЕ состояние питомца, чтобы фронт не собирал его по кускам и не ловил гонки.\n\nstatCapped=true означает: показатель уже был 100 ДО действия, XP не начислен. Фронт всё равно проигрывает анимацию — эмоция бесплатная.\n\ncareXpCapReached=true означает: суточный потолок ухода исчерпан. Кнопки остаются активными, меняется только подпись.",
        "parameters": [ { "$ref": "#/components/parameters/IdempotencyKey" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["kind", "actionId"],
          "properties": {
            "kind": { "$ref": "#/components/schemas/PetActionKind" },
            "actionId": { "type": "string", "format": "uuid",
              "description": "UUID с клиента. Защищает от двойного тапа: повтор возвращает прежний результат." }
          } } } } },
        "responses": {
          "200": { "description": "Питомец обновлён", "content": { "application/json": { "schema": {
            "allOf": [ { "$ref": "#/components/schemas/Envelope" },
              { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ActionResult" } } } ] } } } },
          "409": { "description": "ACTION_LIMIT_REACHED или PET_IS_SLEEPING", "content": { "application/json": { "schema": {
            "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/rewards": {
      "get": {
        "tags": ["rewards"],
        "summary": "Награды",
        "description": "Энтайтлмент, не промокод (решение 10.08, docs/DECISIONS.md): награда — право, привязанное к user_id через reward_grants, коду на экране взяться неоткуда.\n\nconditionLabel и etaDays считает сервер: фронт не склеивает фразы из чисел, иначе смена экономики потребует правки клиента.",
        "parameters": [ { "name": "status", "in": "query", "required": false,
          "schema": { "$ref": "#/components/schemas/RewardStatus" } } ],
        "responses": { "200": { "description": "Список наград", "content": { "application/json": { "schema": {
          "allOf": [ { "$ref": "#/components/schemas/Envelope" },
            { "type": "object", "properties": { "data": { "type": "array",
              "items": { "$ref": "#/components/schemas/Reward" } } } } ] } } } } }
      }
    },
...
    "/rewards/{rewardId}/claim": {
      "post": {
        "tags": ["rewards"],
        "summary": "Забрать награду",
        "description": "Идемпотентно. Повтор с ТЕМ ЖЕ Idempotency-Key возвращает 200 и ту же награду (status уже claimed или used — грант происходит не дважды).\n\nПри другом ключе по уже полученной награде возвращается 409 REWARD_ALREADY_CLAIMED, в data — та же награда, без выдумывания второго гранта — фронт открывает тот же экран, а не показывает ошибку.\n\nЕсли tier=cosmetic, сервер сразу выдаёт предмет и шлёт по сокету cosmetic.granted.",
        "parameters": [
          { "$ref": "#/components/parameters/RewardId" },
          { "$ref": "#/components/parameters/IdempotencyKeyRequired" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/RewardOk" },
          "409": { "description": "REWARD_NOT_ELIGIBLE (условие не выполнено) или REWARD_ALREADY_CLAIMED (в data — та же награда). Оба — состояние, не права доступа, поэтому не 403", "content": { "application/json": { "schema": {
            "allOf": [ { "$ref": "#/components/schemas/ErrorEnvelope" },
              { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Reward" } } } ] } } } },
          "410": { "description": "REWARD_EXPIRED", "content": { "application/json": { "schema": {
            "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/rewards/{rewardId}/redeem-click": {
      "post": {
        "tags": ["rewards"],
        "summary": "Применить награду",
        "description": "В полном продукте — лог перехода «Применить на Авито», статус переходит в used отдельным серверным событием после реального погашения на стороне Авито. В этом MVP такого события нет неоткуда взяться (нет интеграции с настоящим Авито) — статус переходит в used сразу, синхронно; это явное упрощение MVP, а не тихая недоделка (docs/DECISIONS.md → 10.08).",
        "parameters": [ { "$ref": "#/components/parameters/RewardId" } ],
        "responses": {
          "200": { "$ref": "#/components/responses/RewardOk" },
          "409": { "description": "REWARD_NOT_CLAIMED — награду сначала нужно забрать", "content": { "application/json": { "schema": {
            "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key", "in": "header", "required": false,
        "schema": { "type": "string", "format": "uuid" },
        "description": "UUID с клиента. Повтор с тем же ключом возвращает прежний результат, а не ошибку."
      },
      "IdempotencyKeyRequired": {
        "name": "Idempotency-Key", "in": "header", "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Обязателен: ретрай при плохой сети не должен выдать второй грант награды."
      },
...
      "RewardId":   { "name": "rewardId", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "PetOk": {
        "description": "Состояние питомца",
        "content": { "application/json": { "schema": {
          "allOf": [ { "$ref": "#/components/schemas/Envelope" },
            { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Pet" } } } ] } } }
      },
      "RewardOk": {
        "description": "Награда",
        "content": { "application/json": { "schema": {
          "allOf": [ { "$ref": "#/components/schemas/Envelope" },
            { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Reward" } } } ] } } }
      },
      "Unauthorized": {
        "description": "UNAUTHORIZED или TOKEN_EXPIRED. На TOKEN_EXPIRED фронт делает refresh и повторяет запрос один раз",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
      "RateLimited": {
        "description": "ACTION_RATE_LIMITED. В meta.retryAfterSec — через сколько можно повторить",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object", "required": ["meta"],
        "properties": {
          "data": { "description": "Полезная нагрузка. null при ошибке" },
          "meta": { "$ref": "#/components/schemas/Meta" }
        }
      },
      "Meta": {
        "type": "object", "required": ["code", "message"],
        "properties": {
          "code": { "type": "integer", "description": "ВСЕГДА число, дублирует HTTP-статус", "examples": [200] },
          "message": { "type": "string", "description": "Текст для пользователя, на русском" },
          "error": { "$ref": "#/components/schemas/ErrorCode" },
          "requestId": { "type": "string", "format": "uuid",
            "description": "Эхо заголовка X-Request-Id. По нему бэк находит лог по скриншоту бага" },
          "retryAfterSec": { "type": "integer", "description": "Только при 429" }
        }
      },
      "ErrorEnvelope": {
        "type": "object", "required": ["meta"],
        "properties": { "data": { "type": "null" }, "meta": { "$ref": "#/components/schemas/Meta" } }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Машинный код. Фронт ветвится ТОЛЬКО по нему: message перепишут при вычитке текстов, и клиент молча сломается.",
        "enum": [
          "VALIDATION_ERROR", "UNAUTHORIZED", "TOKEN_EXPIRED", "NOT_FOUND",
          "PET_ALREADY_EXISTS", "PET_IS_SLEEPING", "ACTION_LIMIT_REACHED", "ACTION_RATE_LIMITED",
          "EVENT_REJECTED", "REWARD_NOT_ELIGIBLE", "REWARD_ALREADY_CLAIMED", "REWARD_EXPIRED",
          "REWARD_NOT_CLAIMED", "COSMETIC_NOT_OWNED", "INTERNAL_ERROR"
        ]
      },
...
      "PresetId": { "type": "string", "enum": ["green", "blue", "purple"] },
...
      "Stats": {
        "type": "object", "required": ["hunger", "joy", "clean", "energy"],
        "description": "Целые 0..100. Потолок жёсткий: 82 + 30 = 100, остаток сгорает",
        "properties": {
          "hunger": { "type": "integer", "minimum": 0, "maximum": 100 },
          "joy":    { "type": "integer", "minimum": 0, "maximum": 100 },
          "clean":  { "type": "integer", "minimum": 0, "maximum": 100 },
          "energy": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "PetActionKind": { "type": "string", "enum": ["feed", "play", "wash", "sleep", "wake"] },
      "PetMood": {
        "type": "string",
        "description": "Производная от МИНИМАЛЬНОГО показателя. Клиент её не вычисляет",
        "enum": ["radiant", "happy", "neutral", "sad", "sick", "sleeping"]
      },
      "ActionState": {
        "type": "object", "required": ["remaining", "xpCapped"],
        "properties": {
          "remaining": { "type": "integer", "description": "Сколько раз ещё можно сегодня" },
          "cooldownUntil": { "type": "string", "format": "date-time" },
          "xpCapped": { "type": "boolean", "description": "Суточный потолок ухода исчерпан" }
        }
      },
      "Pet": {
        "type": "object",
        "required": ["id", "presetId", "name", "level", "xp", "xpToNext", "totalXp",
                     "stage", "stats", "mood", "moodMultiplier", "actions", "updatedAt"],
        "properties": {
          "id": { "type": "string" },
          "presetId": { "$ref": "#/components/schemas/PresetId" },
          "name": { "type": "string" },
          "level": { "type": "integer", "minimum": 1, "maximum": 25 },
          "xp": { "type": "integer", "description": "XP внутри текущего уровня" },
          "xpToNext": { "type": "integer", "description": "round10(100 * 1.25^(level-1))" },
          "totalXp": { "type": "integer" },
          "stage": { "type": "integer", "enum": [1, 2, 3, 4] },
          "stageLabel": { "type": "string", "examples": ["Искатель"] },
          "stats": { "$ref": "#/components/schemas/Stats" },
          "mood": { "$ref": "#/components/schemas/PetMood" },
          "moodMultiplier": { "type": "number", "examples": [1.0] },
          "streakMultiplier": { "type": "number", "examples": [1.2] },
          "sleepUntil": { "type": "string", "format": "date-time" },
          "actions": { "type": "object", "additionalProperties": { "$ref": "#/components/schemas/ActionState" } },
          "equipped": { "type": "array", "items": { "type": "string" },
            "description": "id надетой косметики" },
          "nextStage": { "type": "object",
            "properties": { "stage": { "type": "integer" }, "label": { "type": "string" },
                            "atLevel": { "type": "integer" } } },
          "updatedAt": { "type": "string", "format": "date-time",
            "description": "Момент, на который посчитан decay. Точка отсчёта для следующего пересчёта" }
        }
      },
      "ActionResult": {
        "type": "object", "required": ["pet", "xpGained", "statCapped", "leveledUp"],
        "properties": {
          "pet": { "$ref": "#/components/schemas/Pet" },
          "xpGained": { "type": "integer", "description": "Уже с множителями и округлением вниз" },
          "statCapped": { "type": "boolean",
            "description": "Показатель был 100 ДО действия. XP не начислен, анимацию всё равно играем" },
          "careXpToday": { "type": "integer", "description": "Сколько XP ухода начислено за сутки" },
          "careXpCapReached": { "type": "boolean" },
          "leveledUp": { "type": "boolean" },
          "unlockedRewardIds": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "RewardTier": {
        "type": "string",
        "description": "cosmetic — себестоимость 0, выдаётся напрямую · soft — бонусы кошелька, мелкие скидки · premium — платные услуги Авито, только с 10 уровня и вехи 100-дневного стрика. premium сегодня не встречается ни в одной награде каталога: вехи 100 без стрик-системы не существует, а выдумывать недостижимую награду нечестнее, чем не показывать её вовсе (docs/DECISIONS.md → 10.08)",
        "enum": ["cosmetic", "soft", "premium"]
      },
      "RewardStatus": { "type": "string", "enum": ["locked", "available", "claimed", "used", "expired"] },
      "Reward": {
        "type": "object",
        "required": ["id", "tier", "title", "conditionLabel", "progress", "status"],
        "description": "Энтайтлмент, не промокод (docs/DECISIONS.md → 10.08): право на бонус привязано к user_id в reward_grants, кода, который можно переслать, в ответе нет вообще.",
        "properties": {
          "id": { "type": "string" },
          "tier": { "$ref": "#/components/schemas/RewardTier" },
          "title": { "type": "string", "examples": ["Скидка на Авито Доставку"] },
          "description": { "type": "string" },
          "conditionLabel": { "type": "string", "examples": ["Уровень 10"],
            "description": "ГОТОВЫЙ текст с сервера. Фронт не склеивает фразы из чисел" },
          "progress": { "type": "object", "required": ["current", "target", "unit"],
            "properties": { "current": { "type": "integer" }, "target": { "type": "integer" },
              "unit": { "type": "string", "enum": ["level", "day", "task"] } } },
          "etaDays": { "type": "integer", "description": "Оценка сервера по темпу пользователя. Отсутствует, если сигнала недостаточно для честной оценки — не выдумывается" },
          "status": { "$ref": "#/components/schemas/RewardStatus" },
          "cosmeticId": { "type": "string", "description": "Если tier=cosmetic" },
          "claimedAt": { "type": "string", "format": "date-time" },
          "usedAt": { "type": "string", "format": "date-time",
            "description": "Когда награда применена (POST /rewards/{rewardId}/redeem-click). Отсутствует, пока status не used" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}